pyproject.toml
[project]
name = "agent-office"
version = "0.1.0"
description = "AI-powered Slack bot for Google Workspace management"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.20.0",
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.150.0",
    "pydantic>=2.9.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "structlog>=24.4.0",
    "aiofiles>=24.1.0",
    "python-magic>=0.4.27",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adapters", "agent", "auth", "core", "services", "tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
